Re-arrange the structure, begin to compose and style components

+++ src/App.js
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
-import Header from './components/Header';
-import Footer from './components/Footer';
-import About from './components/About';
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
+import About from './pages/About';
 import Tasks from './components/Tasks';
 import AddTask from './components/AddTask';
 
@@ -43,8 +43,13 @@ function App() {
   }
 
   const deleteTask = async (id) => {
-    await fetch(`http://localhost:5000/tasks/${id}`, { method: 'DELETE' })
-    setTasks(tasks.filter((task) => task.id !== id))
+    const checker = prompt('Are you sure?', '');
+    if (checker) {
+      await fetch(`http://localhost:5000/tasks/${id}`, { method: 'DELETE' })
+      setTasks(tasks.filter((task) => task.id !== id))
+    } else {
+      return;
+    }
   }
 
   const toggleReminder = async (id) => {
@@ -62,7 +67,7 @@ function App() {
   return (
     <Router>
       <div className='container'>
-        <Header showAddTask={ showAddTask } onAdd={() => setShowAddTask(!showAddTask)} title='Task Manager'/>
+        <Header showAddTask={ showAddTask } onAdd={() => setShowAddTask(!showAddTask)}/>
         <Route exact path='/' render={() => (
           <>
             {
